Move graphql server out of core package
