+++ packages/core/package.json
@@ -8,7 +8,6 @@
     "lodash": "^4.17.11",
     "path": "^0.12.7",
     "pino": "^5.8.0",
-    "graphql-yoga": "1.16.2",
     "fs-extra": "^7.0.0",
 
     "bull": "^3.4.8"
