+++ packages/server/index.js
@@ -1,50 +1,17 @@
-const { server } = require('@openapi/core');
-const Redis = require('ioredis');
+const { GraphQLServer } = require('graphql-yoga');
+
+const { openApi } = require('@openapi/core');
+const log = require('@openapi/core/log').child({ ns: '@openapi/server' });
+
+const config = require('./config');
 
 (async () => {
-  await server.start({
-    db: {
-      redis: {
-        main: new Redis(process.env.OA_REDIS_MAIN_URL, {
-          enableReadyCheck: true,
-        }),
-        queue: new Redis(process.env.OA_REDIS_QUEUE_URL, {
-          enableReadyCheck: true,
-        }),
-      },
-      mongo: {},
-      postgresql: {},
-    },
-    queue: {
-      redisUrl: process.env.OA_REDIS_QUEUE_URL,
-    },
-    enabledProviders: [
-      '@openapi/plugin-common',
-      [
-        '@openapi/plugin-omdb',
-        {
-          auth: {
-            apiKey: process.env.OA_OMDB_API_KEY,
-          },
-        },
-      ],
-      '@openapi/plugin-openlibrary',
-      [
-        '@openapi/plugin-openweather',
-        {
-          auth: {
-            appId: process.env.OA_OPENWEATHER_API_KEY,
-          },
-        },
-      ],
-      [
-        '@openapi/plugin-tvdb',
-        {
-          auth: {
-            apikey: process.env.OA_TVDB_API_KEY,
-          },
-        },
-      ],
-    ],
+  const { typeDefs, resolvers, context } = await openApi.init(config);
+
+  log.info('creating graphql server');
+  const server = new GraphQLServer({ typeDefs, resolvers, context });
+  server.start(() => {
+    // eslint-disable-next-line no-console
+    console.log(`Server is running at http://localhost:4000`);
   });
 })();
