+++ packages/server/package.json
@@ -9,6 +9,7 @@
     "@openapi/plugin-openlibrary": "1.0.0",
     "@openapi/plugin-openweather": "1.0.0",
     "@openapi/plugin-tvdb": "1.0.0",
-    "ioredis": "^4.2.0"
+    "ioredis": "^4.2.0",
+    "graphql-yoga": "1.16.2"
   }
 }
